Extract types and styled components
