+++ editor.planx.uk/src/@planx/components/Feedback/types.ts
@@ -1,0 +1,11 @@
+import { BaseNodeData } from "../shared";
+
+export interface FeedbackComponentProps extends BaseNodeData {
+  title: string;
+  description?: string;
+  fn?: string;
+}
+export interface FormProps {
+  feedbackScore: string;
+  feedback: string;
+}
